Rename interpreter to Datastore.
