+++ src/main/scala/com/ironcorelabs/davenport/syntax/dbprog.scala
@@ -8,7 +8,7 @@ import DB._
 import scalaz.stream.Process
 import scalaz._
 import scalaz.concurrent.Task
-import interpreter.Interpreter
+import datastore.Datastore
 
 // The convention is for syntax objects to start with lower case, so they look
 // like package names. Scalastyle doesn't care for this, so ignore the line.
@@ -17,6 +17,6 @@ final object dbprog extends DBProgOps // scalastyle:ignore
 trait DBProgOps {
   implicit class OurDBProgOps[A](self: DBProg[A]) {
     def process: Process[DBOps, DBError \/ A] = Batch.liftToProcess(self)
-    def interpret(i: Interpreter): Task[DBError \/ A] = i.interpret(self.run)
+    def execute(d: Datastore): Task[DBError \/ A] = d.execute(self.run)
   }
 }
