+++ src/main/scala/com/ironcorelabs/davenport/syntax/process.scala
@@ -8,7 +8,7 @@ import scalaz.concurrent.Task
 import scalaz.stream.Process
 import com.couchbase.client.java.Bucket
 import DB.DBOps
-import interpreter.Interpreter
+import datastore.Datastore
 
 // The convention is for syntax objects to start with lower case, so they look
 // like package names. Scalastyle doesn't care for this, so ignore the line.
@@ -16,7 +16,7 @@ final object process extends ProcessOps // scalastyle:ignore
 
 trait ProcessOps {
   implicit class OurProcessOps[M[_], A](self: Process[M, A]) {
-    def interpret(i: Interpreter)(implicit ev: Process[M, A] =:= Process[DBOps, A]): Process[Task, A] =
-      i.interpretP(self)
+    def execute(d: Datastore)(implicit ev: Process[M, A] =:= Process[DBOps, A]): Process[Task, A] =
+      d.executeP(self)
   }
 }
